backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

google-generativeai==0.2.0
torch==2.1.0
transformers==4.34.0
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.2
joblib==1.3.2

PyMuPDF==1.23.5
easyocr==1.7.1
opencv-python==4.8.1.78
numpy==1.26.1
pandas==2.1.2

python-dotenv==1.0.0
logging==0.4.9.6
pathlib==1.0.1